osd: Store empty range instead of clearing PWLC.
We think this will fix an issue where it is unclear whether the
empty PWLC state is newer or older than a populated PWLC on
another shard.

This is an unfortunate small increase in metadata, so we should
come back to this in a later commit (or possibly later PR).

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>